Extract routing table out of peer manager
