+++ chain/network/src/lib.rs
@@ -8,7 +8,8 @@ mod codec;
 mod peer;
 mod peer_manager;
 pub mod peer_store;
-pub mod types;
 mod rate_counter;
+mod routing;
+pub mod types;
 
 pub mod test_utils;
